Event Attendance (And other Misc. Changes)
The big change is that Event Attendance Manager now exists.
Known bugs which I don't know if I'll be able to fix today
Selecting a project in the project table doesn't update the table below
it to reflect that.
The tables on that page should be sortable/filterable and are not.